Fix panic from write to closed channel
If a backend tries to send a message confirmation following a message
publish, but the server is exiting for any reason, it's possible that
the channel is closed before the write is attempted and we get a panic.

This introduces a test that reproduces the panic, and a fix that makes
the test pass.