Go proximo consumer: wait on WaitGroup last
When an error is returned by the caller supplied function,
`consumeContext` will return that error (after receiving on the errs
channel). In doing so, 3 deferred functions are called, in the following
order (reverse to code order):
* `wg.Wait()`
* `stream.CloseSend()`
* `conn.Close()`

The problem encountered with this is that the first goroutine which is
receiving inbound messages will not stop, and `wg.Wait()` will block
indefinitely.

This change adjust the order of deferred function calls so that they are
invoked as follows:
* `stream.CloseSend()`
* `conn.Close()`
* `wg.Wait()`

This resolves the issue as closing the stream and connection causes the
problematic goroutine to finish. With `wg.Done()` called by both
goroutines, `wg.Wait()` returns, meaning `consumeContext` can return.