Don't block forever in ProducerConn.Produce()
If we can't write to the requests channel, we block forever.  This is a
bug.  There are two scenarios where this happens.  One is when we're
exiting and the context is cancelled, but we don't check for that.  The
second is when we're in error state and nothing is consuming that
channel any longer.

This fixes both of these cases by checking for them in a `select` block.