Don't constrain the message size at the GRPC level
Max message size (if there is one) is a concern of the underlying
broker, so we don't really want a max value at the GRPC level, we want
to delegate it to the broker.  Achieve this here by using large hard
coded values for the GRPC max message size.

There's a question about whether we might want to have a single "max
size" config option that we use for configuring both the broker client
and grpc settings, but not all broker clients have this ability, and its
not clear what benefit that would have anyway.